refactor: removing jentic_agents folder
