+++ memory/base_memory.py
