+++ memory/scratch_pad.py
