+++ tests/test_memory.py
