+++ tools/exceptions.py
